Big changes, finally scraped it all (probably)
So, I finally added all of the ingredients and recipes to my DB, and I thought I would rather sleep on a warm side of my pillow than give all of the ingredients categories by myself. Completely erased IngredientCategoryAdapter because of that. Soon going to create a list with all of the ingredients and a search bar.